chore: fix links in README.md
Links have been broken on the project page (https://peabrainiac.github.io/lean-orbifolds/) for a while now, because the relative links in README.md worked on https://github.com/peabrainiac/lean-orbifolds but not there. The previous commit tried to fix this by changing them to absolute links, but used the wrong absolute URLs; this commit should fix them so the links now work on both versions of the page.